Refactor: move deployment validation logic to separate file.
